Add output format argument
With this change one may choose on the output format with `-o tv`
for tag-value or `-o json` for json. Defaults to `tv`. Includes an
update to the save approach with constructing the final doc object
prior saving it, which is an improvement of the previous
concatenating of data, which cannot construct a proper json output

Signed-off-by: Ivana Atanasova <iyovcheva@vmware.com>